pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypton"
version = "0.1.0"
description = "Automated crypto trading bot using Bollinger Bands and RSI"
authors = [
    {name = "Lazar Gošić", email = "lazar22.gosic2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "ccxt>=4.0.0",
    "python-binance>=1.0.0",
    "pandas>=2.0.0",
    "pandas-ta>=0.3.0",
    "backtrader>=1.9.70",
    "loguru>=0.6.0",
    "python-dotenv>=1.0.0",
    "prometheus-client>=0.14.0",
    "pyyaml>=6.0.0",
    "slack-sdk>=3.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pre-commit>=3.0.0",
    "black>=23.0.0",
    "isort>=5.10.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
markers = [
    "unit: marks test as a unit test",
    "integration: marks test as an integration test",
]
